refactored object drawing code out of app.js
